+++ git_bbb/__init__.py
@@ -2,6 +2,7 @@
 Git-bbb: Brisk Blame Browser for Git.
 """
 from prompt_toolkit import Application
+from prompt_toolkit.enums import EditingMode
 from prompt_toolkit.layout import Layout
 from prompt_toolkit.styles.pygments import style_from_pygments_cls
 from pygments.styles import get_style_by_name
@@ -42,5 +43,7 @@ def run(path, rev, ignore_revs_file):
         mouse_support=True,
         style=style_from_pygments_cls(get_style_by_name(pygments_style)),
     )
+
+    app.editing_mode = EditingMode.VI
 
     app.run()